force two-column output when analysing multiple commits
If we analyse multiple commits for dependencies via the CLI, e.g.

  git deps master~3..master

or

  git deps master~4 master

then we typically want to know not just what the dependencies are,
but which of the input commits caused each dependency.  So use the
same two-column output format which we already used with --recurse.
This also happens to be the same format understood by tsort(1).